Last commit - Live a happy life!
This will be my last commit, for  a lot of uncommitted work of the
beginning of the 2018 year.

For anyone reading this comment, thank you that was a wild trip!
The reason for the end of my role in this adventure can be found on
spigotmc.org :)

Oh and as a quick reminder, otter are really cool animals, you should
definitely take some time to go see them